listening to events from the api, cancellation works

+++ packetblocks-core/src/main/java/org/aincraft/BlockController.java
@@ -2,26 +2,35 @@ package org.aincraft;
 
 import com.google.inject.Inject;
 import io.papermc.paper.event.packet.PlayerChunkLoadEvent;
+import io.papermc.paper.event.packet.PlayerChunkUnloadEvent;
 import java.util.List;
 import java.util.Optional;
 import net.kyori.adventure.key.Key;
+import org.aincraft.BlockBinding.BlockBindingImpl;
 import org.aincraft.PacketBlock.PacketBlockMeta;
+import org.aincraft.events.PacketBlockBreakEvent;
 import org.aincraft.events.PacketBlockInteractEvent;
+import org.aincraft.events.PacketBlockPlaceEvent;
 import org.bukkit.Bukkit;
+import org.bukkit.Chunk;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.BlockExplodeEvent;
+import org.bukkit.event.block.BlockPistonExtendEvent;
+import org.bukkit.event.block.BlockPistonRetractEvent;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.Plugin;
+import org.bukkit.util.Vector;
 import org.jetbrains.annotations.ApiStatus.Internal;
 
 final class BlockController implements Listener {
@@ -31,15 +40,33 @@ final class BlockController implements Listener {
   private final ItemService itemService;
 
   @Inject
-  BlockController(Plugin plugin, PacketBlockService blockService, ItemService itemService) {
+  BlockController(Plugin plugin, PacketBlockService blockService,
+      ItemService itemService) {
     this.plugin = plugin;
     this.blockService = blockService;
     this.itemService = itemService;
   }
 
   @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
-  private void onChunkLoad(final PlayerChunkLoadEvent event) {
+  private void onPlayerChunkLoad(final PlayerChunkLoadEvent event) {
+    Chunk chunk = event.getChunk();
+    Player player = event.getPlayer();
+    List<PacketBlock> blocks = blockService.loadAll(chunk);
+    for (PacketBlock block : blocks) {
+      EntityModel model = block.getModel();
+      model.showTo(player);
+    }
+  }
 
+  @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+  private void onPlayerChunkUnload(final PlayerChunkUnloadEvent event) {
+    Chunk chunk = event.getChunk();
+    Player player = event.getPlayer();
+    List<PacketBlock> blocks = blockService.loadAll(chunk);
+    for (PacketBlock block : blocks) {
+      EntityModel model = block.getModel();
+      model.hideFrom(player);
+    }
   }
 
   @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
@@ -55,32 +82,47 @@ final class BlockController implements Listener {
     }
     Player player = event.getPlayer();
     Bukkit.getPluginManager().callEvent(new PacketBlockInteractEvent(
-        event.getPlayer(), block,
+        player, block,
         event.getAction(), event.getHand(),
         event.getBlockFace(), event.getItem(), packetBlock.getMeta().key().toString())
     );
+    EntityModel model = packetBlock.getModel();
     Bukkit.getScheduler().runTask(plugin, () -> {
-      player.sendBlockChange(location, Bukkit.createBlockData(Material.GLASS));
+      for (Player viewer : model.getViewers()) {
+        viewer.sendBlockChange(location, Bukkit.createBlockData(Material.GLASS));
+      }
     });
   }
 
   @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
   private void onBlockPlace(final BlockPlaceEvent event) {
     ItemStack itemInHand = event.getItemInHand();
-    Optional<String> resourceKey = itemService.read(itemInHand);
-    if (resourceKey.isEmpty()) {
+    Optional<String> data = itemService.read(itemInHand);
+    if (data.isEmpty()) {
       return;
     }
+    String resourceKey = data.get();
     Block block = event.getBlock();
-    PacketBlock packetBlock = blockService.save(
-        BlockBinding.create(block.getLocation(), resourceKey.get()));
-    //TODO: handle for players around the block
-    EntityModel model = packetBlock.getModel();
     Player player = event.getPlayer();
-    model.showTo(player);
+    PacketBlockPlaceEvent blockPlaceEvent = new PacketBlockPlaceEvent(player, block, resourceKey);
+    Bukkit.getPluginManager().callEvent(blockPlaceEvent);
+    if (blockPlaceEvent.isCancelled()) {
+      event.setCancelled(true);
+      return;
+    }
+    Location location = block.getLocation();
+    PacketBlock packetBlock = blockService.save(
+        new BlockBindingImpl(location, resourceKey));
+    EntityModel model = packetBlock.getModel();
+    Chunk chunk = location.getChunk();
+    for (Player viewer : chunk.getPlayersSeeingChunk()) {
+      model.showTo(viewer);
+    }
     Bukkit.getScheduler().runTaskLater(plugin, () -> {
-      player.sendBlockChange(block.getLocation(), Bukkit.createBlockData(Material.GLASS));
-    },2L);
+      for (Player viewer : chunk.getPlayersSeeingChunk()) {
+        viewer.sendBlockChange(block.getLocation(),Bukkit.createBlockData(Material.GLASS));
+      }
+    }, 2L);
   }
 
   @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
@@ -91,9 +133,28 @@ final class BlockController implements Listener {
     if (packetBlock == null) {
       return;
     }
-    blockService.delete(location);
+    Player player = event.getPlayer();
     PacketBlockMeta meta = packetBlock.getMeta();
+    Key key = meta.key();
+    PacketBlockBreakEvent blockBreakEvent = new PacketBlockBreakEvent(player, block,
+        key.toString());
+    Bukkit.getPluginManager().callEvent(blockBreakEvent);
+    if (blockBreakEvent.isCancelled()) {
+      event.setCancelled(true);
+      return;
+    }
+    blockService.delete(location);
     meta.getSoundEntry(SoundType.BREAK).ifPresent(sound -> sound.play(location));
+  }
+
+  @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+  private void onPistonBlockPush(final BlockPistonExtendEvent event) {
+    handlePistonMoveEvent(new PistonEvent(event.getDirection(), event.getBlocks()));
+  }
+
+  @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
+  private void onPistonBlockRetract(final BlockPistonRetractEvent event) {
+    handlePistonMoveEvent(new PistonEvent(event.getDirection(), event.getBlocks()));
   }
 
   @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
@@ -106,15 +167,31 @@ final class BlockController implements Listener {
     handleExplosionEvent(event::blockList);
   }
 
-  private void handleExplosionEvent(final ExplodeEvent event) {
-    List<Block> blocks = event.getBlockList();
-    blocks.forEach(block -> {
+  private void handlePistonMoveEvent(final PistonEvent event) {
+    BlockFace blockFace = event.blockFace();
+    Vector direction = blockFace.getDirection();
+    List<Block> blockList = event.blockList();
+    blockList.forEach(block -> {
       Location location = block.getLocation();
       PacketBlock packetBlock = blockService.load(location);
-      if (packetBlock != null) {
-        blockService.delete(location);
-      }
+      Location newLocation = location.add(direction);
     });
+  }
+
+  private void handleExplosionEvent(final ExplodeEvent event) {
+    for (Block block : event.getBlockList()) {
+      Location location = block.getLocation();
+      PacketBlock packetBlock = blockService.load(location);
+      if (packetBlock == null) {
+        continue;
+      }
+      blockService.delete(location);
+    }
+  }
+
+  @Internal
+  record PistonEvent(BlockFace blockFace, List<Block> blockList) {
+
   }
 
   @Internal
